Move getCourseInnerBlocks helper
