+++ assets/blocks/course-outline/data.js
@@ -1,12 +1,12 @@
 /**
  * External dependencies
  */
-import { invert } from 'lodash';
-
 /**
  * WordPress dependencies
  */
 import { createBlock } from '@wordpress/blocks';
+import { select } from '@wordpress/data';
+import { invert } from 'lodash';
 
 /**
  * Course structure data.
@@ -148,4 +148,25 @@ export const getFirstBlockByName = ( blockName, blocks ) => {
 	}
 
 	return false;
+};
+
+/**
+ * Get the course outline inner blocks of a specific type.
+ *
+ * @param {string} outlineClientId The outline block client id.
+ * @param {string} blockType       The block type to return.
+ *
+ * @return {Array} An array of blocks.
+ */
+export const getCourseInnerBlocks = ( outlineClientId, blockType ) => {
+	let allChildren = select( 'core/block-editor' ).getBlocks(
+		outlineClientId
+	);
+
+	allChildren = allChildren.reduce(
+		( m, block ) => [ ...m, ...block.innerBlocks ],
+		allChildren
+	);
+
+	return allChildren.filter( ( { name } ) => blockType === name );
 };
